[XLA] Fix parameter type of IsLiteralWithValue in algebraic_simplifier.cc.
It should match the param type of LiteralUtil::IsAll.
Change: 146692488

+++ tensorflow/compiler/xla/service/algebraic_simplifier.cc
@@ -46,7 +46,7 @@ namespace xla {
 namespace {
 
 // Returns whether operand is a literal with the given value.
-bool IsLiteralWithValue(const HloInstruction* operand, int value) {
+bool IsLiteralWithValue(const HloInstruction* operand, int8 value) {
   return operand->opcode() == HloOpcode::kConstant &&
          LiteralUtil::IsAll(operand->literal(), value);
 }
